fruitAGNTCY/fruit_agents/recruiter/pyproject.toml
[project]
name = "agent-recruiter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "a2a-sdk==0.3.2",
    "mcp>=1.24.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    # See fruit_cognition/pyproject.toml for rationale on capping <1.83.6.
    "litellm>=1.83.0,<1.83.6",
    "langchain-litellm==0.3.5",
    "langchain-anthropic>=0.3.13",
    "langchain-google-genai>=2.1.4",
    "langchain-openai>=0.3.16",
    "agntcy-app-sdk==0.4.6",
    "google-adk>=1.28.1",
    "rogue-ai-sdk>=0.3.4",
    "pydantic-yaml>=1.6.0",
    # Vulnerability fixes (transitive)
    "langchain-core>=1.2.28",
    "cryptography>=46.0.7",
    "urllib3>=2.6.3",
    "pyasn1>=0.6.3",
    "aiohttp>=3.13.4",
    "nltk>=3.9.4",
    "authlib>=1.6.11",
    "pillow>=12.2.0",
    "orjson>=3.11.6",
    # protobuf: cannot bump to 6.33.5+ — agntcy-oasf-sdk-protocolbuffers-python (dev) pins protobuf>=6.32.1,<6.33.
    # The JSON recursion depth bypass DoS does not affect recruiter (no untrusted protobuf JSON parsing).
    "PyJWT>=2.12.0",
    "python-multipart>=0.0.26",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.0.10",
    "langgraph-checkpoint>=4.0.0",
    "langsmith>=0.7.31",
    "langchain-openai>=1.1.14",
    "Mako>=1.3.11",
    "Pygments>=2.20.0",
    "pypdf>=6.10.2",
]

[project.optional-dependencies]
dev = [
    "agntcy-oasf-sdk-grpc-python==1.75.0.1.20250917120021+8b2bf93bf8dc",
    "agntcy-oasf-sdk-protocolbuffers-python==32.1.0.1.20250917120021+8b2bf93bf8dc",
    "pytest>=9.0.3",
    "pytest-asyncio>=0.24.0",
    "pytest-repeat==0.9.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
url = "https://buf.build/gen/python"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_recruiter"]

[tool.pytest.ini_options]
addopts = "--asyncio-mode=auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "error",
    # google-adk runners.py:1481 reads RunConfig.save_input_blobs_as_artifacts (deprecated).
    # Warning attributed to google.adk.runners. Remove when ADK stops using it.
    "ignore:deprecated:DeprecationWarning:google.adk.runners",
    # google-adk mcp_toolset.py:181 still uses MCPTool (deprecated); should use McpTool. Remove when ADK is updated.
    "ignore:MCPTool class is deprecated:DeprecationWarning:google.adk.tools.mcp_tool.mcp_tool",
]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest-asyncio>=1.3.0"]
